Refactor route implementation into its own file
